requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
astroid==3.3.10
async-timeout==5.0.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
datasets==3.6.0
dill==0.3.8
docker-pycreds==0.4.0
evaluate==0.4.3
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.31.4
idna==3.10
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pylint==3.3.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
transformers==4.52.2
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wandb==0.19.11
xxhash==3.5.0
yarl==1.20.0
